# server/docs/swagger.yaml
basePath: /
definitions:
  api.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  project.CreateRequest:
    properties:
      color:
        type: string
      description:
        type: string
      name:
        type: string
      prefix:
        type: string
    type: object
  project.ListResponse:
    properties:
      projects:
        items:
          $ref: '#/definitions/project.Project'
        type: array
    type: object
  project.Project:
    properties:
      board_id:
        type: string
      color:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      item_seq:
        type: integer
      name:
        type: string
      prefix:
        type: string
    type: object
...
  project.UpdateRequest:
    properties:
      color:
        type: string
      description:
        type: string
      name:
        type: string
    type: object
  queue.CreateRequest:
    properties:
      deadline:
        type: string
      estimate_minutes:
        type: integer
      notes:
        type: string
      priority:
        type: integer
      title:
        type: string
    type: object
  queue.ListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/queue.QueueItem'
        type: array
    type: object
  queue.PromoteRequest:
    properties:
      epic_id:
        type: string
      project_id:
        type: string
      status_id:
        type: string
    type: object
  queue.PromoteResult:
    properties:
      item_display_id:
        type: string
      item_id:
        type: string
    type: object
  queue.QueueItem:
    properties:
      board_id:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      deadline:
        type: string
      estimate_minutes:
        type: integer
      id:
        type: string
      notes:
        type: string
      position:
        type: integer
      priority:
        type: integer
      promoted_item_id:
        type: string
      title:
        type: string
      urgency_score:
        type: number
    type: object
  queue.ReorderRequest:
    properties:
      order:
        items:
          type: string
        type: array
    type: object
  queue.UpdateRequest:
    properties:
      deadline:
        type: string
      estimate_minutes:
        type: integer
      notes:
        type: string
      priority:
        type: integer
      title:
        type: string
    type: object
...
host: api.cumin.ashinsabu.com
info:
  contact: {}
  description: OSS sprint board API
  title: Cumin API
  version: "1.0"
paths:
...
  /api/projects:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/project.ListResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - CookieAuth: []
      summary: List projects
      tags:
      - projects
    post:
      consumes:
      - application/json
      parameters:
      - description: Project to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/project.CreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/project.Project'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Create project
      tags:
      - projects
  /api/projects/{id}:
    delete:
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Delete project
      tags:
      - projects
    patch:
      consumes:
      - application/json
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/project.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/project.Project'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Update project
      tags:
      - projects
  /api/projects/{id}/restore:
    post:
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Restore project
      tags:
      - projects
  /api/queue:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/queue.ListResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - CookieAuth: []
      summary: List queue items
      tags:
      - queue
    post:
      consumes:
      - application/json
      parameters:
      - description: Queue item to capture
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/queue.CreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/queue.QueueItem'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Create queue item
      tags:
      - queue
  /api/queue/{id}:
    delete:
      parameters:
      - description: Queue item ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Archive queue item
      tags:
      - queue
    patch:
      consumes:
      - application/json
      parameters:
      - description: Queue item ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/queue.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/queue.QueueItem'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Update queue item
      tags:
      - queue
  /api/queue/{id}/promote:
    post:
      consumes:
      - application/json
      parameters:
      - description: Queue item ID
        in: path
        name: id
        required: true
        type: string
      - description: Promotion target
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/queue.PromoteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/queue.PromoteResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Promote queue item to board item
      tags:
      - queue
  /api/queue/reorder:
    put:
      consumes:
      - application/json
      parameters:
      - description: Ordered queue item IDs
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/queue.ReorderRequest'
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Reorder queue items
      tags:
      - queue
...
securityDefinitions:
  CookieAuth:
    in: cookie
    name: token
    type: apiKey
swagger: "2.0"